reafactor: change validate api route
