+++ api/pokemonRoutes.ts
@@ -4,7 +4,6 @@ import m from "../data/moves.json";
 import p from "../data/pokemon.json";
 import p2 from "../data/pokemonWithMainSeriesMoves.json";
 import r from "../data/rules.json";
-import onTeamValidate from "../handlers/onTeamValidate";
 
 let moves: any = m;
 let quickmoves: Array<string> = [];
@@ -103,30 +102,6 @@ router.get('/:id', (req, res) => {
         result ? res.json(result) : res.status(404).json(`Could not find Pokemon of id: ${req.params.id}`);
     }catch(err){
         console.error();
-        res.status(500).json({message: "Internal server error"});
-    }
-});
-
-// @desc Validate a Pokemon team
-// @route GET /api/validate/:team/:meta
-// @access Public (for now)
-router.get('/validate/:team/:meta', (req, res) => {
-    try{
-        const meta: any = rules[req.params.meta];
-        if (meta) {
-            const team = JSON.parse(req.params.team)
-            const data = onTeamValidate(team, meta)
-            if (typeof data === 'string') {
-                res.json({
-                    message: data
-                })
-            } else {
-                res.json(data)
-            }
-        } else {
-            res.status(404).json(`${req.params.id} is not a valid meta.`);
-        }
-    } catch(err) {
         res.status(500).json({message: "Internal server error"});
     }
 });
