+++ api/rest_server.ts
@@ -8,6 +8,7 @@ import moveRoutes from "./moveRoutes";
 import userRoutes from "./userRoutes";
 import roomRoutes from "./roomRoutes";
 import ruleRoutes from "./ruleRoutes";
+import validateRoutes from "./validateRoutes";
 
 export const SERVER_PORT = 8081;
 
@@ -34,6 +35,7 @@ app.use('/api/moves', moveRoutes);
 app.use('/api/users', userRoutes);
 app.use('/api/room', roomRoutes);
 app.use('/api/rule', ruleRoutes);
+app.use('/api/validate', validateRoutes);
 
 // serve static files from a given folder
 app.use(e.static('../public'));
